UI: Move OBSBasic service functions in its own file
