+++ UI/cmake/ui-windows.cmake
@@ -20,6 +20,7 @@ target_sources(
           window-basic-main-profiles.cpp
           window-basic-main-scene-collections.cpp
           window-basic-main-screenshot.cpp
+          window-basic-main-service.cpp
           window-basic-main-transitions.cpp
           window-basic-main.cpp
           window-basic-main.hpp
