+++ UI/window-basic-main.cpp
@@ -1302,92 +1302,6 @@ retryScene:
 	}
 }
 
-#define SERVICE_PATH "service.json"
-
-void OBSBasic::SaveService()
-{
-	if (!service)
-		return;
-
-	char serviceJsonPath[512];
-	int ret = GetProfilePath(serviceJsonPath, sizeof(serviceJsonPath),
-				 SERVICE_PATH);
-	if (ret <= 0)
-		return;
-
-	OBSDataAutoRelease data = obs_data_create();
-	OBSDataAutoRelease settings = obs_service_get_settings(service);
-
-	obs_data_set_string(data, "type", obs_service_get_type(service));
-	obs_data_set_obj(data, "settings", settings);
-
-	if (!obs_data_save_json_safe(data, serviceJsonPath, "tmp", "bak"))
-		blog(LOG_WARNING, "Failed to save service");
-}
-
-bool OBSBasic::LoadService()
-{
-	const char *type;
-
-	char serviceJsonPath[512];
-	int ret = GetProfilePath(serviceJsonPath, sizeof(serviceJsonPath),
-				 SERVICE_PATH);
-	if (ret <= 0)
-		return false;
-
-	OBSDataAutoRelease data =
-		obs_data_create_from_json_file_safe(serviceJsonPath, "bak");
-
-	if (!data)
-		return false;
-
-	obs_data_set_default_string(data, "type", "rtmp_common");
-	type = obs_data_get_string(data, "type");
-
-	OBSDataAutoRelease settings = obs_data_get_obj(data, "settings");
-	OBSDataAutoRelease hotkey_data = obs_data_get_obj(data, "hotkeys");
-
-	service = obs_service_create(type, "default_service", settings,
-				     hotkey_data);
-	obs_service_release(service);
-
-	if (!service)
-		return false;
-
-	/* Enforce Opus on FTL if needed */
-	if (strcmp(obs_service_get_protocol(service), "FTL") == 0) {
-		const char *option = config_get_string(
-			basicConfig, "SimpleOutput", "StreamAudioEncoder");
-		if (strcmp(option, "opus") != 0)
-			config_set_string(basicConfig, "SimpleOutput",
-					  "StreamAudioEncoder", "opus");
-
-		option = config_get_string(basicConfig, "AdvOut",
-					   "AudioEncoder");
-		if (strcmp(obs_get_encoder_codec(option), "opus") != 0)
-			config_set_string(basicConfig, "AdvOut", "AudioEncoder",
-					  "ffmpeg_opus");
-	}
-
-	return true;
-}
-
-bool OBSBasic::InitService()
-{
-	ProfileScope("OBSBasic::InitService");
-
-	if (LoadService())
-		return true;
-
-	service = obs_service_create("rtmp_common", "default_service", nullptr,
-				     nullptr);
-	if (!service)
-		return false;
-	obs_service_release(service);
-
-	return true;
-}
-
 static const double scaled_vals[] = {1.0,         1.25, (1.0 / 0.75), 1.5,
 				     (1.0 / 0.6), 1.75, 2.0,          2.25,
 				     2.5,         2.75, 3.0,          0.0};
@@ -4524,23 +4438,6 @@ void OBSBasic::RenderMain(void *data, uint32_t, uint32_t)
 }
 
 /* Main class functions */
-
-obs_service_t *OBSBasic::GetService()
-{
-	if (!service) {
-		service =
-			obs_service_create("rtmp_common", NULL, NULL, nullptr);
-		obs_service_release(service);
-	}
-	return service;
-}
-
-void OBSBasic::SetService(obs_service_t *newService)
-{
-	if (newService) {
-		service = newService;
-	}
-}
 
 int OBSBasic::GetTransitionDuration()
 {
